♻️ firebase.js에서 cart 분리
